moved IDDict
