+++ docs/source/api/classes/xgi.classes.hypergraph.rst
@@ -12,5 +12,3 @@
       :nosignatures:
       
         Hypergraph
-      
-        IDDict
