+++ docs/source/api/utils/xgi.utils.utilities.rst
@@ -11,6 +11,8 @@
       :toctree: .
       :nosignatures:
 
+        IDDict
+
 
    .. rubric:: Functions
