+++ xgi/convert.py
@@ -76,12 +76,9 @@ def convert_to_hypergraph(data, create_using=None):
         H = empty_hypergraph(create_using)
         H.add_nodes_from((n, attr) for n, attr in data.nodes.items())
         ee = data.edges
-        H.add_edges_from(
-            (ee.members(e), e, deepcopy(attr)) for e, attr in ee.items()
-        )
+        H.add_edges_from((ee.members(e), e, deepcopy(attr)) for e, attr in ee.items())
         H._hypergraph = deepcopy(data._hypergraph)
         return H
-    
 
     elif isinstance(data, SimplicialComplex):
         return from_max_simplices(data)
@@ -223,7 +220,7 @@ def convert_to_simplicial_complex(data, create_using=None):
         )
         H._hypergraph = deepcopy(data._hypergraph)
         return H
-    
+
     elif isinstance(data, Hypergraph):
         H = empty_simplicial_complex(create_using)
         H.add_nodes_from((n, attr) for n, attr in data.nodes.items())
@@ -243,7 +240,7 @@ def convert_to_simplicial_complex(data, create_using=None):
         result = from_bipartite_pandas_dataframe(data, create_using)
         if not isinstance(create_using, SimplicialComplex):
             return convert_to_simplicial_complex(result)
-            
+
     elif isinstance(data, dict):
         # edge dict in the form we need
         raise XGIError("Cannot generate SimplicialComplex from simplex dictionary")
